requirements.txt
Flask==3.1.0
python-dotenv==0.0.1 
requests==2.32.3 
Flask-Session==0.8.0
google-generativeai==0.8.4 
gunicorn==23.0.0 